added authentication using JWT to access routes

+++ controller/contactController.js
@@ -8,7 +8,8 @@ const createContact = asyncHandler(async (req,res)=>{
     const createContact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id : req.user.id
     });
     if(createContact){
         res.status(201).json(createContact)
@@ -22,22 +23,24 @@ const createContact = asyncHandler(async (req,res)=>{
 // get
 const getContact = asyncHandler(async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
-    if(contact){
-        res.status(200).json(contact)
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(401);
+        throw new Error('No access to delete');
     }
-    else{
-        res.status(400).json({message : "Something went wrong"})
-    }
+    res.status(200).json(contact)
 });
 
 // update
 const updateContact = asyncHandler(async (req,res)=>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
-        res.status(404).json({message : "Something went wrong"})
+        res.status(401).json({message : "Something went wrong"})
         throw new Error('user not found');
     }
-
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(401);
+        throw new Error('No access to delete');
+    }
     const updateUserContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -55,9 +58,27 @@ const deleteContact = asyncHandler(async (req,res)=>{
         res.status(404);
         throw new Error('user not found');
     }
-    await Contact.deleteOne();
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(404);
+        throw new Error('No access to delete');
+    }
+    await Contact.deleteOne({_id : contact.user_id});
     res.status(201).json(contact);
  });
 
 
- module.exports = { getContact,createContact,updateContact,deleteContact}
+// getAll
+const getAllContacts = asyncHandler(async (req,res)=>{
+    console.log(`req.user.id ${req.user.id}`)
+    const contact = await Contact.find({user_id : req.user.id});
+    if(contact){
+        res.status(200).json(contact)
+    }
+    else{
+        res.status(400).json({message : "Something went wrong"})
+    }
+});
+
+
+ module.exports = { getContact,createContact,updateContact,deleteContact,getAllContacts}
